Split out bubble sort
