+++ src/app.rs
@@ -1,68 +1,5 @@
-use ecolor::Color32;
-use egui::{ecolor, Response, Ui};
-use egui_plot::{Bar, BarChart, Plot};
-use rand::seq::SliceRandom;
-use rand::thread_rng;
-
-pub struct BubbleSort {
-    data: Vec<Bar>,
-    current_loc: usize,
-    changed: bool,
-    finished: bool,
-}
-
-impl BubbleSort {
-    pub fn new(data: Vec<Bar>) -> Self {
-        let mut sort = BubbleSort {
-            data,
-            current_loc: 0,
-            changed: false,
-            finished: false,
-        };
-        sort.data[0].fill = Color32::GREEN;
-        sort
-    }
-
-    pub fn step(&mut self) {
-        // End conditions
-        if self.finished {
-            return;
-        };
-        if self.current_loc + 1 == self.data.len() {
-            if !self.changed {
-                // Nothing more to do
-                self.finished = true;
-                return;
-            }
-            // More work to do, back to the beginning!
-            self.data[self.current_loc].fill = Color32::RED;
-            self.current_loc = 0;
-            self.changed = false;
-        }
-        if self.data[self.current_loc].value > self.data[self.current_loc + 1].value {
-            // Swap them around
-            let temp = self.data[self.current_loc].value;
-            self.data[self.current_loc].value = self.data[self.current_loc + 1].value;
-            self.data[self.current_loc + 1].value = temp;
-            // Note that we've had to make a change
-            self.changed = true;
-        }
-        // Update location
-        self.data[self.current_loc].fill = Color32::RED;
-        self.current_loc += 1;
-        self.data[self.current_loc].fill = Color32::GREEN;
-    }
-
-    // Make this a trait!
-    pub fn plot_chart(&self, ui: &mut Ui) -> Response {
-        let chart = BarChart::new(self.data.clone()).name("Current State");
-        Plot::new("Normal Distribution Demo")
-            .clamp_grid(true)
-            .y_axis_width(3)
-            .show(ui, |plot_ui| plot_ui.bar_chart(chart))
-            .response
-    }
-}
+use crate::bubble::BubbleSort;
+use crate::make_bar_vec;
 
 // TODO: Learn how to disable the state saving.  Don't want it.
 #[derive(serde::Deserialize, serde::Serialize)]
@@ -70,24 +7,6 @@ impl BubbleSort {
 pub struct SortApp {
     #[serde(skip)]
     bubble_sort: BubbleSort,
-}
-
-fn make_bar_vec(size: usize) -> Vec<Bar> {
-    // Produce a randomly shuffled vector of numbers first
-    let mut rng = thread_rng();
-    let mut numbers: Vec<usize> = (0..size).collect();
-    numbers.shuffle(&mut rng);
-
-    // Now turn it in to a vector of BarChart Bars
-    let mut bars: Vec<Bar> = vec![];
-
-    for (index, value) in numbers.iter().enumerate() {
-        let mut bar = Bar::new(index as f64, *value as f64);
-        bar.fill = Color32::RED;
-        bars.push(bar);
-    }
-
-    bars
 }
 
 impl Default for SortApp {
